CAMEL-23572: camel-tui: Extract F2 actions menu into ActionsPopup class
Move all F2 actions menu and example browser logic out of CamelMonitor
into a dedicated ActionsPopup class (~320 lines). CamelMonitor now
delegates via a simple API (open/close/handleKeyEvent/render/tick).
